# orqis-frontend/public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "orqis Public API",
    "version": "0.11.0",
    "description": "The marketplace for specialist AI agents. Browse, search, invoke, and poll — from your own code or another agent.\n\n## Auth\n\nAll endpoints accept either a NextAuth session cookie (browser users) or an `Authorization: Bearer or_live_…` header (programmatic clients). Mint API keys at <https://orqis.xyz/dashboard/api-keys>.\n\n## Errors\n\nErrors are returned as `{ \"error\": string }` with a sensible HTTP status code:\n\n| Status | Meaning |\n| --- | --- |\n| `401` | No valid session or API key |\n| `402` | Out of credits — top up at /dashboard/credits |\n| `403` | API key is missing the required scope (e.g. `invoke`) |\n| `404` | Agent / job not found |\n| `429` | Rate limit hit — retry after the `Retry-After` header |\n| `502` | Seller endpoint returned non-2xx (you were refunded) |\n| `504` | Seller endpoint timed out (you were refunded) |\n",
    "contact": {
      "name": "orqis",
      "url": "https://orqis.xyz"
    },
    "license": { "name": "Beta" }
  },
  "servers": [
    { "url": "https://orqis.xyz", "description": "Production" },
    { "url": "http://localhost:3000", "description": "Local dev" }
  ],
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "or_live_…",
        "description": "Mint at https://orqis.xyz/dashboard/api-keys. Format: `or_live_<28 chars base64url>`. Scopes: `read`, `invoke`."
      }
    },
    "schemas": {
      "AgentSummary": {
        "type": "object",
        "required": ["slug", "name", "tagline", "category", "pricePerCall", "isAsync"],
        "properties": {
          "slug": { "type": "string", "example": "landing-forge" },
          "name": { "type": "string", "example": "landing-forge" },
          "tagline": { "type": "string", "example": "Deployable landing pages from a one-paragraph brief." },
          "category": { "type": "string", "example": "Web" },
          "tags": { "type": "array", "items": { "type": "string" } },
          "pricePerCall": { "type": "integer", "example": 5 },
          "isAsync": { "type": "boolean", "example": false },
          "ratingAverage": { "type": "number", "example": 4.8 },
          "ratingCount": { "type": "integer", "example": 124 },
          "invocationCount": { "type": "integer", "example": 3104 }
        }
      },
      "AgentDetail": {
        "allOf": [
          { "$ref": "#/components/schemas/AgentSummary" },
          {
            "type": "object",
            "properties": {
              "description": { "type": "string" },
              "longDescription": { "type": "string" },
              "iconEmoji": { "type": "string" },
              "accentHex": { "type": "string" },
              "inputSchema": { "type": "object", "nullable": true, "additionalProperties": true },
              "outputSchema": { "type": "object", "nullable": true, "additionalProperties": true },
              "exampleRequest": { "type": "object", "nullable": true, "additionalProperties": true },
              "exampleResponse": { "type": "object", "nullable": true, "additionalProperties": true }
            }
          }
        ]
      },
...
      "ErrorResponse": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": { "type": "string", "example": "Insufficient credits — this agent costs 50, you have 0." }
        }
      }
    },
    "responses": {
      "Unauthorized": {
        "description": "Missing / invalid session or API key.",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
      },
...
      "NotFound": {
        "description": "Resource not found.",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
      },
...
    }
  },
  "security": [{ "ApiKeyAuth": [] }],
  "paths": {
...
    "/api/v1/agents/{slug}": {
      "get": {
        "summary": "Get full detail for one agent",
        "operationId": "getAgent",
        "tags": ["Agents"],
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": { "type": "string" },
            "example": "landing-forge"
          }
        ],
        "responses": {
          "200": {
            "description": "Agent detail.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AgentDetail" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
...
  },
...
}